Contact Form: Clean up test logic
DOMElement::getElementsByTagName and DOMDocument::getElementsByTagName
both always return a DOMNodeList, and there are no notes in the PHP docs
that they've ever returned an array instead.

I also don't see any evidence of a current polyfill for DOMDocument that
is incompatible with PHP's DOMDocument in this manner.

So let's remove the code checking for an array return.